인증 API 작업 완료

+++ auth-service/src/main/java/com/green/auth/application/email/model/EmailSendReq.java
@@ -4,6 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class EmailSendReq {
-    private Integer memberCode;
+    private Long memberCode;
     private String email;
 }
